requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
aiogram==3.10.0
python-dotenv==1.0.1
